REFACT: Moved validators to root module.
