+++ tests/mets_test.py
@@ -30,7 +30,7 @@ from importlib_resources import files
 import tests.resources.xml as XML
 import tests.resources.ips.unpacked as UNPACKED
 
-from ip_validation.infopacks.mets import MetsValidator
+from ip_validation.mets import MetsValidator
 from ip_validation.xml.schema import LOCAL_SCHEMA, get_local_schema
 
 METS_XML = 'METS.xml'
