+++ tests/utils_test.py
@@ -23,7 +23,7 @@
 # under the License.
 #
 """Module that holds common utilities for unit testing."""
-from ip_validation.infopacks.rules import Severity
+from ip_validation.rules import Severity
 def contains_rule_id(error_list, rule_id, severity=Severity.ERROR):
     """Check that a particular error with specified severity is present in a list
     of errors."""
